+++ src/pages/PortfolioPage/lib/mappingPortfolioCard.ts
@@ -1,4 +1,4 @@
-import { TPortfolioCard } from '../constants/type';
+import { type TPortfolioCard } from '../model';
 import { CardPortfolioProps } from '../ui/CardPortfolio/type';
 
 export const mappingPortfolioCard = (