+++ src/pages/PortfolioPage/model/index.ts
@@ -1,2 +1,3 @@
 export { useDragZoom } from './useDragZoom';
 export { useCardPortfolioHandlesMouse } from './useCardPortfolioHandlesMouse';
+export type { TPositionBoard, TPortfolioCard } from './type';