+++ src/pages/PortfolioPage/model/type.ts
@@ -1,0 +1,14 @@
+export type TPositionBoard = {
+	x: number;
+	y: number;
+	rotation: number;
+};
+
+export type TPortfolioCard = {
+	logo: string;
+	title?: string;
+	link: string;
+	settingsView: TPositionBoard & {
+		bgColor: string;
+	};
+};