+++ src/pages/PortfolioPage/ui/ImgPortfolio/index.ts
@@ -1,0 +1,1 @@
+export { ImgPortfolio } from './ImgPortfolio';