+++ src/pages/PortfolioPage/ui/TitleSection/type.ts
@@ -1,0 +1,6 @@
+import { type TPositionBoard } from '../../model';
+
+export type TitleSectionProps = TPositionBoard & {
+	content: string;
+	fontSize: number;
+};